+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Change `torch` dependency from "==2.1.0" to ">=2.1.0,<3.0.0".
 - Change `dataptation` dependency from "==3.1" to ">=3.1,<4.0".
 - Change `optuna` dependency from "3.5.0" to ">=3.5.0,<4.0.0".
+- Add `FourierLayer` implementation.
 
 ## 0.0.0 (2024-02-22)